backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
psycopg2-binary>=2.9
docutils>=0.20
openai>=1.30
gunicorn>=21.2
whitenoise>=6.7
requests>=2.31
dj-database-url>=2.1
